Added new favicon and create checklist link for Index.cshtml, login and register nav buttons now only display when logged out
I have changed the favicon from the DotNet generic to a more personalised one. A new link has been added to index on login to direct new users. Login and register nav buttons are now dependent on the user being logged out in order to display and the other navs only appear when logged in. A small grammar correction made in the privacy policy.